[skip ci] WIP: SARIF compatibility
* Refactoring
